Fix importing of the CSS property database in style-sheets.js
On 2017-09-27 a change landed in mozilla-central that added
a check for SpecialPowers.DOMWindowUtils.isStyledByServo in
property_database.js.  See:

https://hg.mozilla.org/mozilla-central/rev/fd98557b7d2a

This check breaks the importing code in
dom/fuzzer/modules/style-sheets.js since that file's fake
SpecialPowers object doesn't include a DOMWindowUtils
property.  This changeset expands the SpecialPowers object,
as necessary, setting isStyledByServo to true since that's
the default for Firefox now.